Fixed issue where system clock drif/change could result in unexpected behavior
System.currentTimeMillis will be adjusted when the system clock is adjusted.  This may result in timeouts occuring, and other time operations not being accurate.
This uses Threadly's clock to avoid that issue, with a time backed from System.nano so that these system clock changes wont impact our timing.
In addition there was a couple places where we were able to avoid making a clock call at all (there are probably others, but trying to keep it safe for now).