Much better listener thread usage
This makes it so all the listeners are now on a single scheduler.  This allows for much better thread usage since it's unlikely the whole of the connections will all need two threads.
Though we do grow and shrink this scheduler based off how many listeners exist, so that if desired every listener can in fact still use two threads!
There is one possible downside, which is the task queue has the potential to be larger now.  This means that removal of tasks has become more expensive (and we seem to be doing a lot of it).  That said, I don't think it will be impactful, even for large pools.